Update NuiProgress value type.

+++ src/main/Anvil/API/Nui/Widgets/NuiProgress.cs
@@ -10,6 +10,6 @@ namespace Anvil.API
     }
 
     [JsonProperty("value")]
-    public NuiProperty<int> Value { get; set; }
+    public NuiProperty<float> Value { get; set; }
   }
 }
